LayoutTests/http/tests/security/denied-base-url-javascript-url-expected.txt

CONSOLE MESSAGE: Blocked setting javascript:// This is JavaScript as the base URL because it does not have an allowed scheme.

PASS 'javascript:' is an invalid base URL.
